requirements.txt
streamlit
pandas
pypdf2
pdfplumber
python-docx
pillow
pytesseract
scikit-learn
openpyxl